refactor: extract utility function to a separate file
