+++ _website/package.json
@@ -16,6 +16,7 @@
         "react": "^19.0.0",
         "react-dom": "^19.0.0",
         "react-paginate": "^8.3.0",
+        "react-router": "^7.3.0",
         "react-tooltip": "^5.28.0",
         "react-transition-group": "^4.4.5",
         "tailwindcss": "^4.0.6"
