+++ _website/src/App.tsx
@@ -2,11 +2,12 @@ import { useEffect, useMemo, useState } from 'react';
 import Header from './layout/header';
 import List from './components/list/list';
 import Search from './components/search';
-import { titleParser } from './tools/parser/parser';
-import { VersionData } from './tools/parser/types';
+import { titleParser } from './utils/parser/parser';
+import { VersionData } from './utils/parser/types';
 import Pagination from './components/pagination';
-import './App.css';
 import { Tooltip } from 'react-tooltip';
+import { getVersionData } from './utils/getLLPkgstore';
+import './App.css';
 
 function App() {
     const [search, setSearch] = useState('');
@@ -45,17 +46,6 @@ function App() {
             />
         </>
     );
-}
-
-async function getVersionData(): Promise<VersionData> {
-    const response = await fetch('./llpkgstore.json', {
-        method: 'GET',
-        headers: {
-            'Cache-Control': 'no-cache',
-        },
-    });
-    const data = await response.json();
-    return data;
 }
 
 export default App;
