+++ _website/src/components/list/list.tsx
@@ -1,6 +1,6 @@
 import { useState } from 'react';
 import Item from './items';
-import { VersionData } from '../../tools/parser/types';
+import { VersionData } from '../../utils/parser/types';
 import DetailModal from '../detail/detail';
 
 interface ListProps {
